# swagger/swagger_server/models/set_distance_hd_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class SetDistanceHDRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, distance_hd: int=None):  # noqa: E501
        """SetDistanceHDRequest - a model defined in Swagger

        :param distance_hd: The distance_hd of this SetDistanceHDRequest.  # noqa: E501
        :type distance_hd: int
        """
        self.swagger_types = {
            'distance_hd': int
        }

        self.attribute_map = {
            'distance_hd': 'distance_hd'
        }

        self._distance_hd = distance_hd

    @classmethod
    def from_dict(cls, dikt) -> 'SetDistanceHDRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The SetDistanceHDRequest of this SetDistanceHDRequest.  # noqa: E501
        :rtype: SetDistanceHDRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def distance_hd(self) -> int:
        """Gets the distance_hd of this SetDistanceHDRequest.


        :return: The distance_hd of this SetDistanceHDRequest.
        :rtype: int
        """
        return self._distance_hd

    @distance_hd.setter
    def distance_hd(self, distance_hd: int):
        """Sets the distance_hd of this SetDistanceHDRequest.


        :param distance_hd: The distance_hd of this SetDistanceHDRequest.
        :type distance_hd: int
        """
        if distance_hd is None:
            raise ValueError("Invalid value for `distance_hd`, must not be `None`")  # noqa: E501

        self._distance_hd = distance_hd