# swagger/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0"
  title: "mea-ms"
  contact: {}
host: "10.10.10.66:8000"
basePath: "/"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security: []
paths:
...
  /analytics/{id}/video_feed:
    get:
      tags:
      - "Analytics"
      summary: "Generate Frame"
      operationId: "generate_frame"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      security: []
      deprecated: false
      responses:
        "200":
          description: ""
          headers: {}
      x-swagger-router-controller: "swagger_server.controllers.analytics_controller"
  /analytics/{id}/type_analytics:
    get:
      tags:
      - "Analytics"
      summary: "Get Type Analytics"
      operationId: "get_type_analytics"
      produces:
      - "application/json"
      parameters: []
      security: []
      deprecated: false
      responses:
        "200":
          description: ""
          headers: {}
      x-swagger-router-controller: "swagger_server.controllers.analytics_controller"
    post:
      tags:
      - "Analytics"
      summary: "Set Type Analytics"
      operationId: "set_type_analytics"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "Body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/SetTypeAnalyticsRequest"
      security: []
      deprecated: false
      responses:
        "200":
          description: ""
          headers: {}
      x-swagger-router-controller: "swagger_server.controllers.analytics_controller"
...
  /analytics/{pathImage}:
    get:
      tags:
      - "Analytics"
      summary: "Get Image assets"
      operationId: "get_imageassets"
      produces:
      - "application/json"
      parameters:
      - name: "pathImages"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      security: []
      deprecated: false
      responses:
        "200":
          description: ""
          headers: {}
      x-swagger-router-controller: "swagger_server.controllers.analytics_controller"
  /analytics/{id}/polygon:
    get:
      tags:
      - "Analytics"
      summary: "Get Polygon Points"
      operationId: "get_polygon_points"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      security: []
      deprecated: false
      responses:
        "200":
          description: ""
          headers: {}
      x-swagger-router-controller: "swagger_server.controllers.analytics_controller"
    post:
      tags:
      - "Analytics"
      summary: "Set Polygon Points"
      operationId: "set_polygon_points"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "Body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/SetPolygonPointsRequest"
      security: []
      deprecated: false
      responses:
        "200":
          description: ""
          headers: {}
      x-swagger-router-controller: "swagger_server.controllers.analytics_controller"
  /analytics/{id}/distance_hd:
    get:
      tags:
      - "Analytics"
      summary: "Get Distance HD"
      operationId: "get_distance_hd"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      security: []
      deprecated: false
      responses:
        "200":
          description: ""
          headers: {}
      x-swagger-router-controller: "swagger_server.controllers.analytics_controller"
    post:
      tags:
      - "Analytics"
      summary: "Set Distance HD"
      operationId: "set_distance_hd"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "Body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/SetDistanceHDRequest"
      security: []
      deprecated: false
      responses:
        "200":
          description: ""
          headers: {}
      x-swagger-router-controller: "swagger_server.controllers.analytics_controller"
...
securityDefinitions: {}
definitions:
  SetTypeAnalyticsRequest:
    type: "object"
    required:
    - "type_analytics"
    properties:
      type_analytics:
        type: "string"
    title: "SetTypeAnalyticsRequest"
    example:
      type_analytics: "AnalyticsThreeClass"
  SetPolygonPointsRequest:
    type: "object"
    required:
    - "object_perimeter"
    - "outside_secure_area"
    - "polygon"
    properties:
      polygon:
        type: "array"
        items:
          type: "array"
          items:
            type: "integer"
            format: "int32"
      outside_secure_area:
        type: "boolean"
      object_perimeter:
        type: "array"
        items:
          type: "string"
    title: "SetPolygonPointsRequest"
    example:
      polygon:
      - - 0
        - 100
      - - 100
        - 0
      - - 0
        - 200
      - - 100
        - 200
      outside_secure_area: true
      object_perimeter:
      - "Person"
  SetDistanceHDRequest:
    type: "object"
    required:
    - "distance_hd"
    properties:
      distance_hd:
        type: "integer"
        format: "int32"
    title: "SetDistanceHDRequest"
    example:
      distance_hd: 4